mt_q2/requirements.txt
# PyTorch and Deep Learning
torch>=2.0.0
torchtext>=0.15.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
spacy>=3.5.0
datasets>=2.14.0

# Evaluation Metrics
sacrebleu>=2.3.1  # BLEU score
rouge-score>=0.1.2  # ROUGE score
nltk>=3.8  # Additional NLP utilities

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Progress bars and utilities
tqdm>=4.65.0
pyyaml>=6.0

# For Multi30k dataset
torchdata>=0.6.0

# Type hints
typing-extensions>=4.5.0

